moved patch route contents into patch-component
