+++ app/templates/application.hbs
@@ -5,6 +5,6 @@
           {{index}}
         {{/link-to}}</li>{{/each}}
   </ul>
-  <button {{action 'new'}}>Add</button><button {{action 'removeCurrentPatch'}}>Remove</button>
+  <button {{action 'newPatch'}}>Add</button><button {{action 'removeCurrentPatch'}}>Remove</button>
 </header>
 {{outlet}}
