+++ app/templates/components/patch-component.hbs
@@ -1,0 +1,29 @@
+<div class="patch {{newConnectionClass}}">
+
+  <div class="patch-connections">
+    <button {{action 'addModule' 'clock'}}>+Clock</button><button {{action 'addModule' 'out'}}>+Out</button><button {{action 'addModule' 'sequence'}}>+Sequence</button>
+    <p>Source: {{sourceModule.type}} {{sourcePort.label}}</p>
+    <p>Dest: {{destModule.type}} {{destPort.label}}</p>
+    <button {{action 'addConnection' sourcePort destPort}}>Connect</button>
+  </div>
+
+  {{patch-diagram 
+    patch=patch 
+    needsUpdate=diagramNeedsUpdate 
+    needsDraw=diagramNeedsDraw 
+    shouldDrawNewConnection=diagramShouldDrawNewConnection 
+    didUpdate=(action 'diagramDidUpdate')
+    didDraw=(action 'diagramDidDraw')
+  }}
+
+  <div class="modules" id="modules">
+    {{#each patch.modules as |module|}}{{component  module.constructor.modelName
+                                                    module=module 
+                                                    selectPort=(action 'selectModulePort' module) 
+                                                    remove=(action 'removeModule' module) 
+                                                    drawDiagram=(action 'drawDiagram') 
+                                                    setDiagramShouldDrawNewConnectionFrom=(action 'setDiagramShouldDrawNewConnectionFrom')
+          }}{{/each}}
+  </div>
+
+</div>
