+++ app/templates/patch.hbs
@@ -1,28 +1,1 @@
-<div class="patch-connections">
-  <button {{action 'addModule' 'clock'}}>+Clock</button><button {{action 'addModule' 'out'}}>+Out</button><button {{action 'addModule' 'sequence'}}>+Sequence</button>
-  <p>Source: {{sourceModule.type}} {{sourcePort.label}}</p>
-  <p>Dest: {{destModule.type}} {{destPort.label}}</p>
-  <button {{action 'addConnection' sourcePort destPort}}>Connect</button>
-</div>
-
-<div class="patch {{newConnectionClass}}">
-  {{patch-diagram 
-    patch=model 
-    needsUpdate=diagramNeedsUpdate 
-    needsDraw=diagramNeedsDraw 
-    shouldDrawNewConnection=diagramShouldDrawNewConnection 
-    didUpdate=(action 'diagramDidUpdate')
-    didDraw=(action 'diagramDidDraw')
-  }}
-
-  <div class="modules" id="modules">
-    {{#each model.modules as |module|}}{{component  module.constructor.modelName
-                                                    module=module 
-                                                    selectPort=(action 'selectModulePort' module) 
-                                                    remove=(action 'removeModule' module) 
-                                                    drawDiagram=(action 'drawDiagram') 
-                                                    setDiagramShouldDrawNewConnectionFrom=(action 'setDiagramShouldDrawNewConnectionFrom')
-          }}{{/each}}
-  </div>
-
-</div>
+{{patch-component patch=model}}
